Refactor X509 and JWT Sources and Workload API client
