+++ test/workloadapi/test_workload_api_client.py
@@ -20,7 +20,7 @@ from unittest.mock import patch
 import pytest
 
 from pyspiffe.exceptions import ArgumentError
-from pyspiffe.workloadapi.default_workload_api_client import DefaultWorkloadApiClient
+from pyspiffe.workloadapi.workload_api_client import WorkloadApiClient
 
 SPIFFE_SOCKET_ENV = 'SPIFFE_ENDPOINT_SOCKET'
 
@@ -28,22 +28,20 @@ SPIFFE_SOCKET_ENV = 'SPIFFE_ENDPOINT_SOCKET'
 # No SPIFFE_ENDPOINT_SOCKET, and no path passed, raises exception
 def test_instantiate_default_without_var():
     with pytest.raises(ArgumentError) as exception:
-        DefaultWorkloadApiClient(None)
+        WorkloadApiClient(None)
 
     assert (
         str(exception.value)
-        == 'Invalid DefaultWorkloadApiClient configuration: SPIFFE endpoint socket: socket must be set.'
+        == 'Invalid WorkloadApiClient configuration: SPIFFE endpoint socket: socket must be set.'
     )
 
 
 # With SPIFFE_ENDPOINT_SOCKET, and no path passed, succeeds
 def test_instantiate_default_with_var():
     os.environ[SPIFFE_SOCKET_ENV] = 'unix:///tmp/agent.sock'
-    with patch.object(
-        DefaultWorkloadApiClient, '_check_spiffe_socket_exists'
-    ) as mock_check:
+    with patch.object(WorkloadApiClient, '_check_spiffe_socket_exists') as mock_check:
         mock_check.return_value = None
-        wlapi = DefaultWorkloadApiClient(None)
+        wlapi = WorkloadApiClient(None)
         assert wlapi.get_spiffe_endpoint_socket() == 'unix:///tmp/agent.sock'
 
     del os.environ[SPIFFE_SOCKET_ENV]
@@ -51,11 +49,9 @@ def test_instantiate_default_with_var():
 
 # Pass socket path
 def test_instantiate_socket_path():
-    with patch.object(
-        DefaultWorkloadApiClient, '_check_spiffe_socket_exists'
-    ) as mock_check:
+    with patch.object(WorkloadApiClient, '_check_spiffe_socket_exists') as mock_check:
         mock_check.return_value = None
-        wlapi = DefaultWorkloadApiClient(spiffe_socket='unix:///tmp/agent.sock')
+        wlapi = WorkloadApiClient(spiffe_socket='unix:///tmp/agent.sock')
         assert wlapi.get_spiffe_endpoint_socket() == 'unix:///tmp/agent.sock'
 
 
@@ -63,10 +59,10 @@ def test_instantiate_socket_path():
 def test_instantiate_default_with_bad_var():
     os.environ[SPIFFE_SOCKET_ENV] = '/invalid'
     with pytest.raises(ArgumentError) as exception:
-        DefaultWorkloadApiClient(None)
+        WorkloadApiClient(None)
 
     assert (
         str(exception.value)
-        == 'Invalid DefaultWorkloadApiClient configuration: SPIFFE endpoint socket: scheme must be set.'
+        == 'Invalid WorkloadApiClient configuration: SPIFFE endpoint socket: scheme must be set.'
     )
     del os.environ[SPIFFE_SOCKET_ENV]
