+++ test/workloadapi/test_workload_api_client_fetch_x509.py
@@ -26,10 +26,8 @@ from pyspiffe.proto.spiffe import workload_pb2
 from pyspiffe.spiffe_id.spiffe_id import SpiffeId
 from pyspiffe.spiffe_id.spiffe_id import TrustDomain
 from pyspiffe.workloadapi.exceptions import FetchX509SvidError, FetchX509BundleError
+from pyspiffe.workloadapi.workload_api_client import WorkloadApiClient
 
-from src.pyspiffe.workloadapi.default_workload_api_client import (
-    DefaultWorkloadApiClient,
-)
 from test.utils.utils import (
     ResponseHolder,
     handle_success,
@@ -50,11 +48,9 @@ from test.workloadapi.test_constants import (
 
 @pytest.fixture
 def client():
-    with patch.object(
-        DefaultWorkloadApiClient, '_check_spiffe_socket_exists'
-    ) as mock_check:
+    with patch.object(WorkloadApiClient, '_check_spiffe_socket_exists') as mock_check:
         mock_check.return_value = None
-        client_instance = DefaultWorkloadApiClient('unix:///dummy.path')
+        client_instance = WorkloadApiClient('unix:///dummy.path')
     return client_instance
